+++ sklearn/ensemble/_base.py
@@ -11,6 +11,7 @@ from joblib import effective_n_jobs
 
 from ..base import BaseEstimator, MetaEstimatorMixin, clone, is_classifier, is_regressor
 from ..utils import Bunch, _print_elapsed_time, check_random_state
+from ..utils._tags import _safe_tags
 from ..utils.metaestimators import _BaseComposition
 
 
@@ -285,3 +286,16 @@ class _BaseHeterogeneousEnsemble(
             names mapped to their values.
         """
         return super()._get_params("estimators", deep=deep)
+
+    def _more_tags(self):
+        try:
+            allow_nan = all(
+                _safe_tags(est[1])["allow_nan"] if est[1] != "drop" else True
+                for est in self.estimators
+            )
+        except Exception:
+            # If `estimators` does not comply with our API (list of tuples) then it will
+            # fail. In this case, we assume that `allow_nan` is False but the parameter
+            # validation will raise an error during `fit`.
+            allow_nan = False
+        return {"preserves_dtype": [], "allow_nan": allow_nan}
