+++ sklearn/ensemble/_voting.py
@@ -30,7 +30,6 @@ from ..preprocessing import LabelEncoder
 from ..utils import Bunch
 from ..utils._estimator_html_repr import _VisualBlock
 from ..utils._param_validation import StrOptions
-from ..utils._tags import _safe_tags
 from ..utils.metadata_routing import (
     _raise_for_unsupported_routing,
     _RoutingNotSupportedMixin,
@@ -152,19 +151,6 @@ class _BaseVoting(TransformerMixin, _BaseHeterogeneousEnsemble):
     def _sk_visual_block_(self):
         names, estimators = zip(*self.estimators)
         return _VisualBlock("parallel", estimators, names=names)
-
-    def _more_tags(self):
-        try:
-            allow_nan = all(
-                _safe_tags(est[1])["allow_nan"] if est[1] != "drop" else True
-                for est in self.estimators
-            )
-        except Exception:
-            # If `estimators` does not comply with our API (list of tuples) then it will
-            # fail. In this case, we assume that `allow_nan` is False but the parameter
-            # validation will raise an error during `fit`.
-            allow_nan = False
-        return {"preserves_dtype": [], "allow_nan": allow_nan}
 
 
 class VotingClassifier(_RoutingNotSupportedMixin, ClassifierMixin, _BaseVoting):
